Translated a lot of code from Flashpunk
I'm using a Flashpunk game as a starting point/test bed for building up the engine. The path of least resistance, then, is to make Atomic have the same interface as Flashpunk.

There are a few places where I've deviated to do my own thing, and quite a lot not implemented. Over time it should feel less Flashpunky.